Pass RGB555 colors from LCD controller to display output
Before, we passed the (DMG) 2-bit greyscale values from
the LCD controller to the display and the display did the
translation to output colors.

Now, the LCD controller does the translation when in
DMG mode (for now: always) and passes RGB555 colors to
the display output, so when later the CGB LCD features
are implemented, we can actually display colors.

Truecolor on the terminal is actually quite slow, so
for now, translate the colors to the 256-color ANSI
palette for the terminal output driver. The Sixel
driver supports the full 15-bit palette.

Truecolor for terminal will be added later.

+++ src/display/display.rs
@@ -1,5 +1,5 @@
-/// Type of a color definition.
-pub type Color = u32;
+/// Type of a color definition (RGB555)
+pub type Color = u16;
 
 /// Base trait for a display output
 pub trait Display: std::any::Any {
